// react-frontend/src/api/models/finance-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TransactionDTO } from './transaction-dto';

/**
 * 
 * @export
 * @interface FinanceUserDTO
 */
export interface FinanceUserDTO {
    /**
     * 
     * @type {number}
     * @memberof FinanceUserDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof FinanceUserDTO
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof FinanceUserDTO
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof FinanceUserDTO
     */
    'lastName'?: string;
    /**
     * 
     * @type {string}
     * @memberof FinanceUserDTO
     */
    'credentialNumber'?: string;
    /**
     * 
     * @type {number}
     * @memberof FinanceUserDTO
     */
    'balance'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinanceUserDTO
     */
    'movementDay'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinanceUserDTO
     */
    'movementWeek'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinanceUserDTO
     */
    'movementMonth'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinanceUserDTO
     */
    'movementYTD'?: number;
    /**
     * 
     * @type {Array<TransactionDTO>}
     * @memberof FinanceUserDTO
     */
    'transactions'?: Array<TransactionDTO>;
}